+++ src/lib.rs
@@ -78,9 +78,10 @@
 //! [`server::handshake`]: server/fn.handshake.html
 //! [`client::handshake`]: client/fn.handshake.html
 
-#![doc(html_root_url = "https://docs.rs/h2/0.3.15")]
+#![doc(html_root_url = "https://docs.rs/h2/0.3.16")]
 #![deny(missing_debug_implementations, missing_docs)]
 #![cfg_attr(test, deny(warnings))]
+#![allow(clippy::type_complexity, clippy::manual_range_contains)]
 
 macro_rules! proto_err {
     (conn: $($msg:tt)+) => {