+++ src/proto/streams/prioritize.rs
@@ -7,9 +7,11 @@ use crate::codec::UserError;
 use crate::codec::UserError::*;
 
 use bytes::buf::{Buf, Take};
-use std::io;
-use std::task::{Context, Poll, Waker};
-use std::{cmp, fmt, mem};
+use std::{
+    cmp::{self, Ordering},
+    fmt, io, mem,
+    task::{Context, Poll, Waker},
+};
 
 /// # Warning
 ///
@@ -235,39 +237,43 @@ impl Prioritize {
         // If it were less, then we could never send out the buffered data.
         let capacity = (capacity as usize) + stream.buffered_send_data;
 
-        if capacity == stream.requested_send_capacity as usize {
-            // Nothing to do
-        } else if capacity < stream.requested_send_capacity as usize {
-            // Update the target requested capacity
-            stream.requested_send_capacity = capacity as WindowSize;
-
-            // Currently available capacity assigned to the stream
-            let available = stream.send_flow.available().as_size();
-
-            // If the stream has more assigned capacity than requested, reclaim
-            // some for the connection
-            if available as usize > capacity {
-                let diff = available - capacity as WindowSize;
-
-                stream.send_flow.claim_capacity(diff);
-
-                self.assign_connection_capacity(diff, stream, counts);
+        match capacity.cmp(&(stream.requested_send_capacity as usize)) {
+            Ordering::Equal => {
+                // Nothing to do
             }
-        } else {
-            // If trying to *add* capacity, but the stream send side is closed,
-            // there's nothing to be done.
-            if stream.state.is_send_closed() {
-                return;
+            Ordering::Less => {
+                // Update the target requested capacity
+                stream.requested_send_capacity = capacity as WindowSize;
+
+                // Currently available capacity assigned to the stream
+                let available = stream.send_flow.available().as_size();
+
+                // If the stream has more assigned capacity than requested, reclaim
+                // some for the connection
+                if available as usize > capacity {
+                    let diff = available - capacity as WindowSize;
+
+                    stream.send_flow.claim_capacity(diff);
+
+                    self.assign_connection_capacity(diff, stream, counts);
+                }
             }
+            Ordering::Greater => {
+                // If trying to *add* capacity, but the stream send side is closed,
+                // there's nothing to be done.
+                if stream.state.is_send_closed() {
+                    return;
+                }
 
-            // Update the target requested capacity
-            stream.requested_send_capacity =
-                cmp::min(capacity, WindowSize::MAX as usize) as WindowSize;
+                // Update the target requested capacity
+                stream.requested_send_capacity =
+                    cmp::min(capacity, WindowSize::MAX as usize) as WindowSize;
 
-            // Try to assign additional capacity to the stream. If none is
-            // currently available, the stream will be queued to receive some
-            // when more becomes available.
-            self.try_assign_capacity(stream);
+                // Try to assign additional capacity to the stream. If none is
+                // currently available, the stream will be queued to receive some
+                // when more becomes available.
+                self.try_assign_capacity(stream);
+            }
         }
     }
 
@@ -317,9 +323,11 @@ impl Prioritize {
     /// connection
     pub fn reclaim_all_capacity(&mut self, stream: &mut store::Ptr, counts: &mut Counts) {
         let available = stream.send_flow.available().as_size();
-        stream.send_flow.claim_capacity(available);
-        // Re-assign all capacity to the connection
-        self.assign_connection_capacity(available, stream, counts);
+        if available > 0 {
+            stream.send_flow.claim_capacity(available);
+            // Re-assign all capacity to the connection
+            self.assign_connection_capacity(available, stream, counts);
+        }
     }
 
     /// Reclaim just reserved capacity, not buffered capacity, and re-assign
@@ -372,11 +380,11 @@ impl Prioritize {
                 continue;
             }
 
-            counts.transition(stream, |_, mut stream| {
+            counts.transition(stream, |_, stream| {
                 // Try to assign capacity to the stream. This will also re-queue the
                 // stream if there isn't enough connection level capacity to fulfill
                 // the capacity request.
-                self.try_assign_capacity(&mut stream);
+                self.try_assign_capacity(stream);
             })
         }
     }
@@ -738,21 +746,19 @@ impl Prioritize {
                             // capacity at this point.
                             debug_assert!(len <= self.flow.window_size());
 
+                            // Check if the stream level window the peer knows is available. In some
+                            // scenarios, maybe the window we know is available but the window which
+                            // peer knows is not.
+                            if len > 0 && len > stream.send_flow.window_size() {
+                                stream.pending_send.push_front(buffer, frame.into());
+                                continue;
+                            }
+
                             tracing::trace!(len, "sending data frame");
 
                             // Update the flow control
                             tracing::trace_span!("updating stream flow").in_scope(|| {
-                                stream.send_flow.send_data(len);
-
-                                // Decrement the stream's buffered data counter
-                                debug_assert!(stream.buffered_send_data >= len as usize);
-                                stream.buffered_send_data -= len as usize;
-                                stream.requested_send_capacity -= len;
-
-                                // If the capacity was limited because of the
-                                // max_send_buffer_size, then consider waking
-                                // the send task again...
-                                stream.notify_if_can_buffer_more(self.max_buffer_size);
+                                stream.send_data(len, self.max_buffer_size);
 
                                 // Assign the capacity back to the connection that
                                 // was just consumed from the stream in the previous