+++ src/proto/streams/store.rs
@@ -1,7 +1,5 @@
 use super::*;
 
-use slab;
-
 use indexmap::{self, IndexMap};
 
 use std::convert::Infallible;
@@ -302,15 +300,15 @@ where
             let mut stream = store.resolve(idxs.head);
 
             if idxs.head == idxs.tail {
-                assert!(N::next(&*stream).is_none());
+                assert!(N::next(&stream).is_none());
                 self.indices = None;
             } else {
-                idxs.head = N::take_next(&mut *stream).unwrap();
+                idxs.head = N::take_next(&mut stream).unwrap();
                 self.indices = Some(idxs);
             }
 
-            debug_assert!(N::is_queued(&*stream));
-            N::set_queued(&mut *stream, false);
+            debug_assert!(N::is_queued(&stream));
+            N::set_queued(&mut stream, false);
 
             return Some(stream);
         }
@@ -347,7 +345,7 @@ impl<'a> Ptr<'a> {
     }
 
     pub fn store_mut(&mut self) -> &mut Store {
-        &mut self.store
+        self.store
     }
 
     /// Remove the stream from the store