+++ src/proto/streams/streams.rs
@@ -311,6 +311,7 @@ where
 impl<B> DynStreams<'_, B> {
     pub fn recv_headers(&mut self, frame: frame::Headers) -> Result<(), Error> {
         let mut me = self.inner.lock().unwrap();
+
         if me.store.num_wired_streams() > me.counts.max_streams() {
             tracing::error!(
                 "HEADERS: number of streams exceeds the upper limit ({})",
@@ -321,29 +322,30 @@ impl<B> DynStreams<'_, B> {
                 Reason::PROTOCOL_ERROR,
             ));
         }
-        me.recv_headers(self.peer, &self.send_buffer, frame)
+
+        me.recv_headers(self.peer, self.send_buffer, frame)
     }
 
     pub fn recv_data(&mut self, frame: frame::Data) -> Result<(), Error> {
         let mut me = self.inner.lock().unwrap();
-        me.recv_data(self.peer, &self.send_buffer, frame)
+        me.recv_data(self.peer, self.send_buffer, frame)
     }
 
     pub fn recv_reset(&mut self, frame: frame::Reset) -> Result<(), Error> {
         let mut me = self.inner.lock().unwrap();
 
-        me.recv_reset(&self.send_buffer, frame)
+        me.recv_reset(self.send_buffer, frame)
     }
 
     /// Notify all streams that a connection-level error happened.
     pub fn handle_error(&mut self, err: proto::Error) -> StreamId {
         let mut me = self.inner.lock().unwrap();
-        me.handle_error(&self.send_buffer, err)
+        me.handle_error(self.send_buffer, err)
     }
 
     pub fn recv_go_away(&mut self, frame: &frame::GoAway) -> Result<(), Error> {
         let mut me = self.inner.lock().unwrap();
-        me.recv_go_away(&self.send_buffer, frame)
+        me.recv_go_away(self.send_buffer, frame)
     }
 
     pub fn last_processed_id(&self) -> StreamId {
@@ -352,7 +354,7 @@ impl<B> DynStreams<'_, B> {
 
     pub fn recv_window_update(&mut self, frame: frame::WindowUpdate) -> Result<(), Error> {
         let mut me = self.inner.lock().unwrap();
-        me.recv_window_update(&self.send_buffer, frame)
+        me.recv_window_update(self.send_buffer, frame)
     }
 
     pub fn recv_push_promise(&mut self, frame: frame::PushPromise) -> Result<(), Error> {
@@ -368,17 +370,18 @@ impl<B> DynStreams<'_, B> {
                 Reason::PROTOCOL_ERROR,
             ));
         }
-        me.recv_push_promise(&self.send_buffer, frame)
+
+        me.recv_push_promise(self.send_buffer, frame)
     }
 
     pub fn recv_eof(&mut self, clear_pending_accept: bool) -> Result<(), ()> {
         let mut me = self.inner.lock().map_err(|_| ())?;
-        me.recv_eof(&self.send_buffer, clear_pending_accept)
+        me.recv_eof(self.send_buffer, clear_pending_accept)
     }
 
     pub fn send_reset(&mut self, id: StreamId, reason: Reason) {
         let mut me = self.inner.lock().unwrap();
-        me.send_reset(&self.send_buffer, id, reason)
+        me.send_reset(self.send_buffer, id, reason)
     }
 
     pub fn send_go_away(&mut self, last_processed_id: StreamId) {
@@ -745,7 +748,7 @@ impl Inner {
             }
             None => {
                 proto_err!(conn: "recv_push_promise: initiating stream is in an invalid state");
-                return Err(Error::library_go_away(Reason::PROTOCOL_ERROR).into());
+                return Err(Error::library_go_away(Reason::PROTOCOL_ERROR));
             }
         };
 
@@ -826,7 +829,13 @@ impl Inner {
         let send_buffer = &mut *send_buffer;
 
         if actions.conn_error.is_none() {
-            actions.conn_error = Some(io::Error::from(io::ErrorKind::BrokenPipe).into());
+            actions.conn_error = Some(
+                io::Error::new(
+                    io::ErrorKind::BrokenPipe,
+                    "connection closed because of a broken pipe",
+                )
+                .into(),
+            );
         }
 
         tracing::trace!("Streams::recv_eof");
@@ -1166,7 +1175,7 @@ impl<B> StreamRef<B> {
             let mut child_stream = me.store.resolve(child_key);
             child_stream.unlink();
             child_stream.remove();
-            return Err(err.into());
+            return Err(err);
         }
 
         me.refs += 1;
@@ -1249,10 +1258,7 @@ impl<B> StreamRef<B> {
             .map_err(From::from)
     }
 
-    pub fn clone_to_opaque(&self) -> OpaqueStreamRef
-    where
-        B: 'static,
-    {
+    pub fn clone_to_opaque(&self) -> OpaqueStreamRef {
         self.opaque.clone()
     }
 
@@ -1365,12 +1371,13 @@ impl OpaqueStreamRef {
             .release_capacity(capacity, &mut stream, &mut me.actions.task)
     }
 
+    /// Clear the receive queue and set the status to no longer receive data frames.
     pub(crate) fn clear_recv_buffer(&mut self) {
         let mut me = self.inner.lock().unwrap();
         let me = &mut *me;
 
         let mut stream = me.store.resolve(self.key);
-
+        stream.is_recv = false;
         me.actions.recv.clear_recv_buffer(&mut stream);
     }
 
@@ -1412,7 +1419,7 @@ impl Clone for OpaqueStreamRef {
 
         OpaqueStreamRef {
             inner: self.inner.clone(),
-            key: self.key.clone(),
+            key: self.key,
         }
     }
 }