+++ tests/h2-support/src/prelude.rs
@@ -90,7 +90,7 @@ pub trait ClientExt {
 impl<T, B> ClientExt for client::Connection<T, B>
 where
     T: AsyncRead + AsyncWrite + Unpin + 'static,
-    B: Buf + Unpin + 'static,
+    B: Buf,
 {
     fn run<'a, F: Future + Unpin + 'a>(
         &'a mut self,
@@ -103,7 +103,7 @@ where
                     // Connection is done...
                     b.await
                 }
-                Right((v, _)) => return v,
+                Right((v, _)) => v,
                 Left((Err(e), _)) => panic!("err: {:?}", e),
             }
         })